Kapt: support correcting delegated property error type in K2
In K2, the delegate expression is used as the source of the property,
instead of the encompassing "by ..." expression. So we support both K1
and K2 cases in JvmIrDeclarationOrigin.

Another way to fix this would be to replicate K1 behavior in
`PsiRawFirBuilder.Visitor.toFirProperty`, however that change affects
far more places in diagnostic reporting and it isn't obvious that it is
better, while this fix is local.

 #KT-69860