[JVM] Ignore some tests on local variables with IR inliner
In these tests we generate one more line number. For example, see code
below

```
/* 1 */ try {
/* 2 */     // ... some code ...
/* 3 */     return b
/* 4 */ } finally {
/* 5 */     x = "OK"
/* 6 */ }
/* 7 */ // ...
```

With bytecode inliner we generate line numbers in order 1, 2, 3, 5, 7.
With Ir inliner lines are: 1, 2, 3, 5, 3, 7. There is an additional
line number "3". This probably happens because with bytecode inliner
there is an additional optimization pass on inlined code that can
eliminate unnecessary instructions.