[K/Wasm] Don't add mappings into source-maps for unavailable sources
IDEA JavaScript debugger gets stuck if many files are listed inside source maps (^FL-27262).
To help the debugger, we filter sources from modules not in the user project.
It's a temporary solution until the third-party libraries debugging is supported (it depends on the issue with file absolute path in KLIB by default ^KT-58406)