[K/N] Widen thread_id to uintptr_t
Currently inside the K/N runtime thread id is assumed to be 32-bit.

But in practice it's not always the case. So, when thread id provided
by `pthread_threadid_np` on iOS doesn't fit into 32 bits, runtime 
crashes. We've seen this in practice.

This commit makes thread id pointer size instead, and adjusts the
related code in runtime and compiler. Having it pointer size is
practically enough to guarantee that atomic operations on it will
still be possible.

^KT-70365