[FIR2IR] Resolve FIR for Fir2LazyProperty when creating backing field
The unresolved FIR for a Fir2LazyProperty can cause issues like
unresolved value argument mapping for an annoation constructor for the
property. For example,
```
class EpisodeToPodcast {
    @Relation(parentColumn = "podcast_uri", entityColumn = "uri")
    lateinit var _podcasts: List<String>
...
```
We met an exception caused by the above code on IDE when the compose
compiler plugin accesses the `Fir2IrLazyClass::declarations` in an IR
lowering transformer. The property is one of declarations in the
`Fir2IrLazyClass`, and it calls an annotation constructor with
arguments. The unresolved `FirAnnotationCall` results in a wrong value
argument mapping.

This commit fixes it by running `lazyResolveToPhase(..)` on FIR for
`Fir2LazyProperty` when creating its backing field.

^KT-70545