Resolve containing declaration of annotation call on convertToIrConstructorCall
The compose compiler plugin accesses the `Fir2IrLazyClass::declarations`
in an IR lowering transformer. When a declaration in `Fir2IrLazyClass`
calls an annotation constructor with arguments like a property in the
following code:
```
class EpisodeToPodcast {
    @Relation(parentColumn = "podcast_uri", entityColumn = "uri")
    lateinit var _podcasts: List<String>
...
```
unresolved `FirAnnotationCall` results in a wrong value argument
mapping. This commit fixes it by running `lazyResolveToPhase(..)` for
the containing declaration (it is Fir2LazyProperty for the above
example) of the annoation constructor call on
`convertToIrConstructorCall(..)`.

^KT-70545