[Analysis API] forbid `analyze` in dumb mode
The Analysis API is not designed and supposed to work in the dumb mode.
Due to not fully initialized project, we may have problems with broken
caches and inconsistent states.
For instance, in K2, to analyze something, we have to create a resolution
session which requires some information during initialization and uses
it for the entire `KaSession` lifetime. Hense, if the session was
configured incorrectly due to the dumb mode, then all further `analyze`
will fail even in the smart mode.
Also, some logic under the hood already has some assertions that throw
exceptions in the case of dumb mode.

^KT-63490 Fixed