[AA] Pre-resolve Java use-site member scope to STATUS
The problem was:
- During processing functions, use-site member scope gets overrides
- This leads to the creation of substitution overrides
- Check FirFakeOverrideGenerator#checkStatusIsResolved fails

Q: Why not lazy resolve to STATUS during collecting overrides? E.g. in
`FirClassSubstitutionScope#createSubstitutionOverrideProperty`, where
lazy resolve to TYPES currently happens.
A: Resolving to STATUS from scopes code leads to contract violations.

Q: Why the problem happens only in Java, and not in Kotlin?
A: In case of `declaredMemberScope`, Kotlin doesn't calculate overrides.
In case of `memberScope`, members are pre-resolved to STATUS.

^KT-69070 Fixed