K2: Resolve safe call receiver in ReceiverResolution mode
For safe calls under PCLA, we should behave in the same way as
for regular qualified calls, i.e. we should semi-fix type variables
when necessary

In PCLAInferenceSession semi-fixation only runs for ReceiverResolution.

Then, it's necessary to apply integerLiteralAndOperatorApproximationTransformer
because by default ReceiverResolution leaves them not-approximated,
but constants/ILT evaluation didn't work before and is not supposed
to work for safe calls.

^KT-69170 Fixed