CLI: rename -Xuse-kapt4 to -Xuse-k2-kapt
The name "kapt4" isn't going to make a lot of sense after we reimplement
K2 kapt with the same approach that is used in K1 kapt, aka "kapt3"
(KT-69860).

This argument is not supposed to be used manually by users (one would
use either a Gradle property or a `-Kapt-use-k2` CLI flag), so it's
supposed to be not a user-visible change, but a minor cleanup.