[Analysis API] Fix "No classifier found" exception for using nested `PsiClsClass` in `LLFirFirClassByPsiClassProvider`
For such classes, the separator used between nested classes is `$`
instead of `.`. Symbol providers expect the `.` separator.
We cannot use `PsiClass.qualifiedName.replace('$', '.')`
 as `$` is a valid identifier in Java classes.
A test for such a case is added.

It's complicated to add a test for `PsiClsClass` in the Kotlin
repository, so it's added in the IntelliJ repository.

KT-69416