[Analysis API] Fix "No classifier found" exception for using local `PsiClsClass` in `LLFirFirClassByPsiClassProvider`.
`PsiClsClass.qualifiedName` for such classes looks something like
`javax.swing.JSlider$1SmartHashtable.LabelUIResource`.

`LLFirFirClassByPsiClassProvider` does not work for local classes,
so they are filtered out.

They are also filtered out in `PsiClass.namedClassSymbol` in
`KaFirJavaInteroperabilityComponent` to avoid creating
such `KaFirPsiJavaClassSymbol`s.

It's complicated to add a test for `PsiClsClass` in the Kotlin
repository, so it's added in the IntelliJ repository.

^KT-69416 fixed