KT-67483 [kotlin] Introduce `LLFirSwitchableExtensionDeclarationsSymbolProvider`
Open up `FirSwitchableExtensionDeclarationsSymbolProvider`, use a simple
boolean there for the CLI mode of the compiler.

Use `ThreadLocal` implementation of the `disabled` flag to the
`LLFirSwitchableExtensionDeclarationsSymbolProvider` inheritor.
Create it for all LL sessions.

This change allows the provider to be safely enabled or disabled on a
per-thread basis during lazy resolution in IDE mode.