KT-67483 [kotlin] Disable `FirSwitchableExtensionDeclarationsSymbolProvider` during lazy `COMPILER_REQUIRED_ANNOTATIONS`
This allows to avoid an SOE from the compiler plugins, namely from
the Serialization compiler plugin.

An SOE happened because of the resolve chain like this:

- LLFirCompilerRequiredAnnotationsTargetResolver (1) ->
- FirSwitchableExtensionDeclarationsSymbolProvider ->
- FirExtensionDeclarationsSymbolProvider ->
- SerializationFirResolveExtension ->
- resolvedCompilerRequiredAnnotations ->
- LLFirCompilerRequiredAnnotationsTargetResolver (2) -> ...

The problem did not arise in the CLI mode,
because `FirSwitchableExtensionDeclarationsSymbolProvider` was
always disabled there during the `COMPILER_REQUIRED_ANNOTATIONS` phase
to avoid the interference with compiler plugins.

Now it is also disabled in the corresponding lazy transformer in the
IDE mode.

Note 1: To properly test that this change indeed mitigates
the SOE problem, we had to introduce
the `resolvedCompilerAnnotationsWithClassIds`
call to the test `MembersOfSerializerGenerator` compiler plugin.

Note 2: To avoid the problems with multiple nested calls of
`withGeneratedDeclarationsSymbolProviderDisabled`, `isDisabled`
flag from `FirSwitchableExtensionDeclarationsSymbolProvider` is
explicitly checked before disabling it.
Also, assertions were added to avoid accidents with the premature
enabling.

^KT-67483 Fixed