KT-70852: Introduce swift attributes in SIR
Instead of making SirAttribute a part of SIR nodes, for now we create
a sealed hierarchy of classes which represent a specific Swift attribute

Motivation: it is a bit simpler from implementation point of view
as the number of possible attributes is limited at the moment.
We can always lift this restriction later.