[FIR2IR] Always pre-approximate parameter types of callable reference type
The logic was previously only applied to reflect types and only manually
handled intersection types.
In fact, approximation does the same and more.
The reason why we need to pre-approximate the parameter types is that
we want to approximate them as to their supertypes.
When the whole type is approximated during the call to `toIrType()`,
the type arguments might be approximated differently than top-level
types.

#KT-70625 Fixed