[JVM] Optimize MaxStackFrameSizeAndLocalsCalculator
The algorithm already assumes that every incoming edge to a basic
block has the same stack size, so we can use this assumption to further
simplify the code. For example, it's unnecessary to track a set of
visited blocks because we can mark them with negative input stack sizes;
also, tracking exception handlers is also redundant, as it only fixes
the input stack size at 1 which only has to be done once.