[K/N][tests] Enforce two-stage mode for stdlib tests
Stdlib tests rely on `-Xmulti-platform` by using expect/actual in test
utils. One-stage mode is incompatible with `-Xmulti-platform` for K2.
So, running K/N tests in one-stage mode with K2 fails because of this.

One-stage mode is not a priority, and it is acceptable to have a smaller
test suite there. Given that, the most practical solution for the
problem is to always run stdlib tests in two-stage mode, even if the
test run specifies one-stage.

This commit does exactly that.

Alternatively, we could just ignore stdlib tests in that case, but it
seems less preferable: CI test runs may use one-stage in combination
with other non-default parameters. So by ignoring the tests we would
lose test coverage for those as well.

^KT-70538 Fixed