[K/N][tests] Increase timeouts for stdlib tests from 4 to 10 minutes
Sometimes on slower hosts tests seem to take extremely long time.
This might be caused by swapping or heavy parallel workload. This
affects mostly macOS arm64.

Now, after KT-65593, we see this for stdlib tests with aggressive GC
scheduler and thread state checker enabled (which are expected to be
long).

This commit bumps the timeout for K/N stdlib tests from 4 to 10 minutes.
It also matches the default timeout for codegen/box tests.