[K/N][tests] Don't compare `@ObjCMethod` for 32-bit targets
Certain tests that run cinterop and compare the contents of the
resulting klibs with the golden data.
In particular, they compare the values in `@ObjCMethod` annotations,
e.g. `encoding` value which contains the Objective-C method encoding.
This value depends on the pointer size, and the golden data is in fact
64-bit-specific.

So such tests fail on targets with 32-bit pointers.

To address that, this commit disables the test if it attempts to compare
klib contents with the golden data that has `@ObjCMethod` annotations,
if the test target has 32-bit pointers.

^KT-70980