[K/N][tests] Fix KonanDriverTests in compileOnly mode
A pair of tests in `KonanDriverTests` not only runs the resulting
executable but also compares its output. This doesn't work when
`compileOnly=true` parameter is specified: the execution result is fake
in that case, and the "output" is empty. The corresponding tests fail.

This commit fixes that by muting the parts of the tests comparing output
if `NoOpExecutor` is used.