[FIR] SignatureEnhancement: avoid resolution under lock
The problem is that previously `performFirstRoundOfBoundsResolution`
was executed under the `typeParameterBoundsResolveLock` lock which
is potentially dangerous as the method calls `resolveIfJavaType`,
which may jump into other declarations.
This commit extracts the computation logic out of the lock, so only
publication should be done under the lock.
`performSecondRoundOfBoundsResolution` has been worked similarly.

^KT-70327