[FIR] Java: introduce unified entry point for enhanced annotations
In this way it is much clearer where we have lazy annotations.
Also, this helps us to avoid capturing of use-site sessions.
This should help with memory a bit as we won't hold
annotations PSI directly in lambdas. Annotation owners
are held on smart pointers in the Analysis API mode.
This commit also adds lazy annotation calculation for
`FirJavaTypeParameter`.

^KT-70327