[K/N][tests] Bump timeout in stdlib `WorkerTest.waitForMultipleFutures`
In some test runs on the CI, the tests seem to run longer and have
unexpected delay. For example, when the agent is overloaded, and/or
the test runs with the 'aggressive' GC scheduler.

WorkerTest.waitForMultipleFutures expects to get results from futures
through `waitForMultipleFutures` in 10 seconds and fails if it takes
more.
The test fails now every time in the new test configuration:
'aggressive' scheduler with thread state checker on Linux.

This commit fixes that by setting timeout for `waitForMultipleFutures`
to 20 minutes instead of 10 seconds.
This shouldn't affect the regular test run duration, since the function
exits as soon as it gets the first future result and won't wait for 20
minutes every time.