[JS] Move JS AST optimizer tests to test tasks related to the IR backend
These tests test the functionality still used by the IR backend
(see the usages of the `optimizeFragmentByJsAst` function).

But they were only run as part of the legacy backend-related test task,
which we're about to drop. To continue running these tests, we move them
to the IR backend-related test tasks.

KT-70362