[FIR] Fix implicit visibility of setter overriding a val in redundancy check
For determining the implicit visibility of a setter of a property,
only the visibilities of overridden setters were previously considered.
This included synthetic setters, which are generated for vars.
If the overridden property was a val however, all visibilities were
ignored and a default (private) was used.
This commit considers the visibility of the property itself if the
property has no setter.

^KT-69190 Fixed