[JS] Fix JS IC issue caused by two different files with the same name
The issue initially manifested itself during an incremental build
of the Kotlin/JS standard library using the K2 compiler.

So, when compiling the K/JS stdlib, there are two files:
- `libraries/stdlib/src/kotlin/Any.kt` — this one contains
  `expect class Any`;
- `libraries/stdlib/build/src/js-builtin-sources/libraries/stdlib/jvm/builtins/Any.kt`
  — this one contains `actual class Any`.

After clean-building the stdlib, the Kotlin metadata of both those files
is (correctly) stored in the IC cache.

After we change a file in the stdlib that makes use of Any
(so, just about any file), and try to incrementally build the stdlib,
the following funny thing happens.

In `FirCachedSymbolNamesProvider`, we eventually call
`MetadataLibraryBasedSymbolProvider#knownTopLevelClassesInPackage`,
where we build the set of top level classes inside a package by
iterating over strings that `MetadataLibrary#packageMetadataParts`
returns in `MetadataLibraryBasedSymbolProvider#forEachFragmentInPackage`

Then those strings are eventually fed to
`MetadataLibrary#packageMetadata`, which should return the byte array
of the serialized metadata for that package part.

In the case of JS IC, the implementor of the `MetadataLibrary` interface
that we deal with is `KlibIcData`. It uses source file names as keys
in its iternal map, and those file names are what
`KlibIcData#packageMetadataParts` returns and what
`KlibIcData#packageMetadata` expects.

HOWEVER.

The internal map `KlibIcData#parts` is built from another map —
`incrementalData.compiledPackageParts`, which happens to be a hash map.
Of course when iterating over a hash map, we can expect its entries to
come in arbitrary order.

And because we use file names (not paths!) as keys, if we have two files
with the same name (like `Any.kt`), we might just first put the file
with `actual class Any` to the map we're building, and some time
after put the file with `expect class Any` to that map. The latter then
overwrites the former.

Expect declarations are interesting, because that their IC metadata is
empty, so the symbol provider happily ignores such declaration, and we
end up without a symbol for `Any`. Which is, as you may expect,
a catastrophe.

We fix this by using full paths to files as keys in the map,
rather than only their names. It is safe because this only
concerns incremental compilation.

^KT-71121 Fixed