[FIR] Remove extra condition for reporting MANY_IMPL_MEMBER_NOT_IMPLEMENTED in delegation
In KT-56720, an extra condition was
added to prevent MANY_IMPL_MEMBER_NOT_IMPLEMENTED from being reported
when the same Java method was delegated to multiple times.
The fix was probably incorrect because the Java method in question
was default and shouldn't have been delegated to at all.
This was later fixed by KT-62334.

#KT-71215 Fixed