[Analysis API] Ignore DFA produced by the element itself
Before, the 'ContextCollector' searched for the last 'CFGNode' produced
by an element. This was incorrect, as smart casts produced by the
statement do not affect the statement itself; they affect the following
statements.

Now, smart casts for certain elements (such as 'foo?.bar' safe calls)
is available in the 'ContextKind.BODY' mode.

^KT-71259 Fixed