Make @SubclassOptInRequired accept multiple experimental markers
[3/5] Drop SubclassOptInRequired in KaModule.kt

The TeamCity build for KT-70562 failed because the IDE is using
kotlin-stdlib version 1.9.25 (which doesn't include the vararg
changes), and kotlin was built with a version that includes these
changes.

To fix the issue, the SubclassOptInRequired annotation was temporarily
removed from the affected Kotlin classes.

Once the kotlin-stdlib version in IntelliJ is updated, the
SubclassOptInRequired annotation will be added back in KT-71152.