Inference: Consider fork points when checking if CS has constraints
In this particular forkPointCausedByWarnAnnotation.kt test, the fork
point is a bit trivial: B<String, Boolean> has two versions
of java.util.function.Function supertypes
(one of them has an enhanced-warnings attribute)

And after that, once we conditionally add Unit constraint to the expected
type position, it succeeds since information from the fork point
which contributes information that Vv should be Boolean is not handled
yet.

And only at the completion stage, once we've got both Unit and Boolean
constraints, it reports contradiction.

That idea is to fix `hasConstraint` to make it check that current
fork points are still resolvable.

Also, take a look at `.latestLV` changes which became much cleaner.

^KT-68768 Fixed