K2: don't count code fragment as a declaration in the diagnostic visitor
This commit is a follow-up to KT-70891 fix.
Originally it was supposed, that code fragments are important to be
in a checker context containing declaration list.
In fact, it's not so, as we can find a code fragment in
a containing file declaration list.
Moreover, the code like
'checkerContext.containingDeclarations.last() is FirSomething'
is used in a bunch of checkers, and it can become broken
if we add a code fragment to the end of the list.
So in this commit we decided to drop code fragments from
checker context containing declarations, effectively reverting
a piece of KT-70891 fix.