[KLIB Resolver] Don't filter libraries passed via Native CLI
It does not make sense to filter the list of libraries passed via CLI
parameter '-library' aka '-l' to remove items which don't have
file separator characters. This list of libraries is anyway passed
as "directLibs" to the search path resolver, which ignores any items
that do not look like absolute paths.

So, by removing this check, we don't change the behavior but simplify
the library resolution in Native CLI.

^KT-67448