[KLIB Resolver] Report error if a library was passed via CLI by unique name
Note: The chosen approach is a kind of hack: Instead of doing honest
check inside KLIB resolver if a unique name was passed as one of
user-defined libs (e.g. via `-library` or `-l` CLI argument), we do
post-check. The post-check runs after the KLIB resolver, and just
makes sure that every item that has been passed as a user-defined lib
and looks like a unique name actually matches the file name of the KLIB.

This approach works good enough. And it is much simpler to implement
than implementing the "honest" check. Anyway, it's a temporary
solution that will exist until we replace the KLIB resolved by
a completely new, much simpler and well-tested component.

Note: This check may give false positives in case unique name
includes path separator character. We consider this as an ultimately
rare case, and accept the possible risks.

^KT-71455 Fixed
^KT-67448 Fixed