[JS] Use JUnit tags for filtering tests i/o manually specifying classes
Two tags are introduced for non-default compilation modes:
`legacy-frontend` and `es6`.

This will make sure that if a new test class is added, we don't have to
not forget to add it to the build script explicitly.
If the newly added test class doesn't have any tags, it will be
run as part of the K2 + ES5 test task (`:js:js.tests:jsFirTest`).