[IR] Introduce new IR parameter API
`parameters` is a new way to store and access value parameters in
IrFunction. It is just a flat list of all kinds of parameters, where
the exact kind (e.g. dispatch receiver, regular parameter) is designated
by `IrValueParameter.kind`.

It is meant to replace the old API, namely:
- valueParameters
- contextReceiverParametersCount
- dispatchReceiverParameter, extensionReceiverParameter

The old API is still there, but implemented atop of `parameters`.

Migration to the new API will follow.

#KT-71582 Fixed