Tests: refactor defaultArguments/reflection tests (1/2)
These tests would be more convenient as normal box tests because then
they'd be run on every combination of compiler frontend/backend and
additional features that we have.

In this commit, I've only changed the content of test data files, so
that git would recognize the move in the next commit.

 #KT-71197 Fixed