[PSI] prefer greenStub to stub
In many cases, we can get information from stubs more efficiently
as it contains major information.
For instance, `KtClassOrObject#isLocal` in stubs is just a boolean flag,
but for AST it is a big PSI check, which requires parents traversal.
The same for `KtClassLikeDeclaration#getClassId` – in stub we already
have `ClassId`, but for AST we need to calculate it

^KT-71565 Fixed