[FIR] Optimize scope of enum entry anonymous object without declarations
While trivial enum entries of enums like `enum class Foo { A, B }` don't
have initializers, entries of enums with constructors like
`enum class Foo(val v: Int) { A0(0) }` do.
Every enum entry will have an anonymous object as the initializer for
which a scope will be built (at the latest during checkers phase).
The scope of a class contains a cache with the callable names that
is a superset of the callable names of the supertypes.
In the case of an enum entry, the callable names contain every
enum entry's name.
As a result, we get quadratic memory consumption (scaling with the
number of enum entries).
This commit optimizes the case where the enum entry doesn't declare any
members by just returning the enum's scope.
The case with enum entries declaring members is not optimized
(see #KT-71591).

The execution time of
compiler/testData/diagnostics/tests/enum/giantEnumWithConstructor.kt
is reduced from 15 s to 7 s.

#KT-71159 Fixed