[IR] Add more function shape parameters to IrCallImplWithShape
and IrFunctionReferenceWithShape.

For now, the provided values are ignored.
However, they will be required for the new argument API - more precisely,
for the bridge between old and new argument APIs.
The old one distinguishes between receiver and regular arguments at
call side with separate properties, while the
new one only has a single list of all arguments.
Hence the distinguishing information needs to be preserved for a case
when argument is added via new API, and later read by the old one.

KT-70054