[FIR] When subjects with dynamic types should be considered Any?
Right now, for when exhaustiveness checking, the lower-bound of flexible
types is chosen. Dynamic types are flexible with a lower-bound of
Nothing. This leads to the subject type being Nothing and no branches
are required to be considered exhaustive. Dynamic types should instead
use their upper-bound (Any?) for exhaustive checks.

^KT-71601 Fixed