Native: Split Objective-C export tests into 3 groups
For historical reasons, Objective-C export tests are implemented using
a separate testing infrastructure which executes all tests in a single
batch. This behaviour might lead to CI timeouts on highly loaded agents.
Proper refactoring will require a lot of work: for example, it is not
obvious how to avoid recompiling stdlib in the release mode. Instead,
we split existing tests into 3 groups which seems like an ok'ish starting
point for the future improvements.