[K/N] Use SpinLock for GlobalData::init() on Windows
There is a weird bug on Windows when sometimes GlobalData has started
initialization and created the timer thread for the GC scheduler;
but when the timer thread tries to wait for the GlobalData
initialization to finish, it crashes on trying to lock the
synchronization mutex. GlobalData initialization does successfully
take the synchronization mutex, so the mutex should definitely be
initialized.

Replace std::mutex with SpinLock on Windows to workaround the problem.

^KT-71644