[LL FIR] Make dangling file see private declarations of its original
...file during visibility checking.

When analyzing a dangling file (e.g. in case of completion) we may
resolve to declaration in the original module in case of
`KaDanglingFileResolutionMode.IGNORE_SELF` mode. This leads to the
problem KT-70356 – false INVISIBLE_REFERENCE errors and private members
not getting into completion.

Q: Why not analyze in different resolution mode - PREFER_SELF?
A: It could be a nice fix, because we would resolve to
declarations in same dangling file, so `==` check would work.
But it leads to another problem – public declarations duplicate
(one from the original module, one from dangling).
This is a known problem, and it could be fixed by introducing a new
resolution mode (see KT-68929).

^KT-70356