[Gradle] Configure mandatory attributes of the non-packed klib artifact
This commit could be just a one-liner similar to
```
artifact.classifier = null
```
Due to some bug fixed in Gradle 8.4, the tasks registered as artifact-producing, were instantiated even if the actual artifacts are not required. Setting the `classifier`, in turn, helps to avoid the instantiation.
Additionally, the things became even worse if the resolution runs at execution time and there are usages of `TaskContainer.names`, because implementation of the latter internally differs instantiated and non-instantiated tasks, potentially causing ConcurrentModificationException during combining them.
That spent almost a working day of 3 engineers to debug the problem :sweat-smile:
^KT-71328 Verification Pending