Include sources in kotlin-osgi-bundle
With this change, we include the sources of all dependencies of
`kotlin-osgi-bundle` in its source bundle.
This gives users of this library a better IDE and debugging experience
providing documentation and allows users to debug in the stdlib.

This implementation works very similarly on how the class files of the
dependencies are merged. This also means that currently conflicts are
not handled at all.
This is probably fine, since there should actually be no duplicates.

^KT-28492 Fixed