[Gradle] Make KotlinSecondaryVariantsDataSharing lenient to duplicates
In some wierd cases. Like declaring two android Targets with different
name. Will cause registration of two "TargetCoordinates".
And previous version of KSVDS were strict on that regard.

I decided to soften the requirement on the task presence and in case
if secondary variant already registered, we report strong warning.
But still lenient.

^KT-57280