[Ir.Actualizer] Fix property to field actualization
In before, the link to temporary field inside f/o property
was left in tree. In fact, this doesn't break anything in
normal compilation, because of the backend ability to work
with slightly incorrect IR.

But it can become a problem after some stdlib code changes,
or when -Xserialize-ir is enabled.

^KT-71826 Fixed