[JS] Split `ES6ConstructorLowering` into two separate lowerings
These lowerings are
1. `ES6SyntheticPrimaryConstructorLowering` that lowers only
synthetic primary constructors
2. `ES6ConstructorLowering` that lowers everything else.

We need this separation because during the second lowering we can
stumble upon synthetic constructor. Earlier we generated a new
call on this constructor with a new extension receiver and transformed
it into an `init` call later on. This is wrong because the original
constructor doesn't have an extension receiver parameter. To fix it,
we can generate an `init` call right away.

#KT-71855 Fixed