CLI: uninherit VirtualFileFinderFactory from MetadataFinderFactory
Use composition instead. The reason is that VirtualFileFinderFactory,
which is used from debugger, is declared in 'frontend.java' and thus is
packed into the `kotlin-compiler-common` artifact for IDE, but
MetadataFinderFactory is in 'frontend', and is located in the
`kotlin-compiler-fe10` artifact only.

It's already incorrect and confusing that `kotlin-compiler-common`
depends on 'frontend.java' but not 'frontend', but that is a separate
problem (there's a TODO in `val commonCompilerModules` about it).

This change only makes it so that at least clients of
VirtualFileFinderFactory would not need to depend on
`kotlin-compiler-fe10` just to be able to resolve its supertype
MetadataFinderFactory (which is not used anywhere from intellij).

 #KT-70139