JVM: move/remove descriptions of phases
`AbstractNamedCompilerPhase.description` is not used anymore since the
previous commit, so in this change JVM (and some common) phase
descriptions are mostly moved to kdocs on lowering phase classes where
applicable, or removed if they were very short and did not provide any
additional context.

Note that for common lowering phases which are inherited in different
backends, the kdoc is usually on the superclass, unless the behavior in
the subclass is different (e.g. see `JvmInitializersCleanupLowering`).

 #KT-71944