K/N: move/remove descriptions of phases
`AbstractNamedCompilerPhase.description` is not used anymore since a few
commits ago, and phase descriptions are mostly moved to kdocs on
lowering phase classes where applicable.

In case of most K/N phases, the descriptions were very short and did not
provide any additional context, so I've just removed them. In case a
phase was from the common backend and kdoc was already there, I've also
removed the description. In cases when descriptions contained some
useful info, I've moved them to the lowering phase class, or the
variable declaring the phase.

Note that for common lowering phases which are inherited in different
backends, the kdoc is usually on the superclass, unless the behavior in
the subclass is different.

 #KT-71944