[FIR] propagate parameter annotations to data class copy properly
The problem with the previous implementations is that `owner` of
recreated parameter annotations was the containing class, which is
incorrect in this case as we have to transform the `copy` method
itself to resolve such annotations.
There are two possible solutions:
1. (Easiest) Fix the owner for annotations
2. Avoid annotation recreation completely to reduce the number of FIR
tree nodes which should be resolved separately

I choose the second one. In this case, we completely avoid recreation,
but a copy of the `FirCallAnnotation` is still necessary if we need to
drop `useSiteTarget`

^KT-71348 Fixed