IR: fix rendering of class references in KotlinLikeDumper
`IrClassReference.symbol` is the class only, but `classType` gives the
full type, which is important in case of `kotlin.Array`. Also, in some
cases `symbol` is incorrect, e.g. see `kClassReference` in
`JvmIrUtils.kt` where the symbol is always `KClass`.

This is not a user-visible change because KotlinLikeDumper is only used
in tests.