[K/N][tests] Fix the type of argument passed to NSInteger parameter
The function `handleFailureInFunction` in `NSAssertionHandler` class
of the Foundation library has the 3rd parameter of type `NSInteger`.
This could be either `kotlin.Long` or `kotlin.Int` depending
on the concrete target.

The fix passes the 3rd argument as `kotlin.Int` assuming that in
case `NSInteger` is `kotlin.Long`, the compiler will take care
of automatic type coercion.