[FIR] Describe how to enable an `UNSUPPORTED_FEATURE`
It seems there's no nice way to have a single source of truth for these
flags: `CommonCompilerArguments` depends on the current module and the
names declared in a `HashMap` wouldn't be accessible inside the
`@Argument` annotations. Adding the names directly to `LanguageFeature`
seems a bit wrong, LFs are something more abstract that doesn't
necessarily require having a CLI name.

It's also possible to move `@Argument`, `@Enables` and `@Disables` to
a separate module that the current one would depend on (as moving them
here feels wrong), but it seems like too much hassle and we're only
using the map built at runtime to make some diagnostic messages a bit
prettier.

^KT-61272 Fixed