[Tests] Allow NativeEnvironmentConfigurator to fix targets per test
It is necessary to be able to fix a certain Native target for a certain
test, and give this target the preference over the target name that is
passed by the user (or by the CI server) via the
'kotlin.internal.native.test.target' parameter.

To achieve this, a new test directive `WITH_FIXED_TARGET` is introduced.

^KT-71333