[Daemon] Combine daemon startup with the client registration
To avoid daemon shutdowns due to delays on initial connection to the daemon, we can register the client immediately during the startup, given the client provides the registration data. This way, even if client is slow (e.g. to high system load at the time), the daemon would still register the client properly.
^KT-69929 Fixed