[Stubs] Fix ClassId for object literals
Before, the 'ClassId' for object literals was '/<no name provided>'
which does not make a lot of sense. Also, it was inconsistent with the
'ClassIdCalculator', so `KtClassOrObject.getClassId()` worked
differently, depending on whether a stub is available or not.

The problem caused object literals to be treated as 'lazy resolvable'
sometimes (see implementation of 'declarationCanBeLazilyResolved()').

^KT-71741 Fixed