PCLA: Allow leaving type variables in the result for semi-fixation
Currently, once we've got two lower bounds, like:
- Nothing? <: OTv
- GenericKlass<GKTv> <: OTv

When computing, their CST, it results in GenericKlass<GKTv>,
but for non-PCLA cases (non-semi-fixation), it's illegal to leave
type variables (stub types) in the fixation result, so the idea for
that case is filtering them out.

In that particular case, it leads to quite an irrelevant `Nothing?`
result being inferred.

For PCLA, we had a part where we allow leaving type variables once we
don't have "proper-proper" (without stub types inside) options.

And the idea of the fix is leaving it for semi-fixation unconditionally
(even if there are "proper-proper" constraints).

^KT-72116 Fixed