[PSI] KtEnumEntry: provide correct `isEquivalentTo`
After 9980a65e7ddb2efd5ec78f02a4b7d36c9be007cb `isEquivalentTo` has
been changed for `KtEnumEntry` as well. And the problem is that the new
implementation depends on `ClassId,` which is always `null` for entries.
This commit fixes this problem. The logic is too specific for the
IntelliJ plugin, so the tests are written there.

^KTIJ-31588 Fixed
^KT-72142 Fixed