[FIR] Fix recursive upper bound check in SAM conversion
When we substitute star projections in SAM types with the parameter's
upper bound, we check if the bound contains a type parameter type
itself, so that we can substitute these recursively.
This commit fixes the check in the case that the argument was not a star
projection but a type parameter type of the SAM type.
This fixes a false negative type mismatch.

#KT-71321 Fixed