KotlinCliJavaFileManagerImpl: don't cache empty search result
The result wasn't cached if findVirtualFileForTopLevelClass failed prior
to 23e5f77ff, which makes the empty list cached. AnalysisHandlers like
KSP rely on the previous behavior to introduce new Java files in
repeated analyses.

This patch reverts to the old behavior with minimal changes.

A better fix might be exposing the ability to let callers of
KotlinCliJavaFileManager invalidate the cache but that would be a larger
change.