[FIR] Fix capturing of captured types
Before this commit, capturing of captured types used the helper function
`ConeCapturedType.substitute` that has a problematic early return
when neither `lowerType` nor `constructor.projection` need
substitution (see #KT-64024).
This means that for a type `CapturedType<*>` with a supertype
`Foo<*>` (from an upper bound of the type parameter), no capturing
of the supertype would be performed.
This commit introduces a new function similar to
`ConeCapturedType.substitute` but without the problematic early return.
This fixes a false-positive ARGUMENT_TYPE_MISMATCH.

#KT-72177 Fixed