[K/N] Remove return slot for object returns
This remove the extra argument for the return slot for functions that
return a memory managed object, thereby transfering the responsibility
(of adding the returned objects to the roots) to the caller of the
function.

This change is to make it more transparent to LLVM where local values
are written to, potentially allowing better optimizations when newer GC
support features of LLVM are used.