IR serializer: Generate stable local signatures only for fake overrides
In fact, stable signatures are only necessary for fake overrides.
This is because it's a separate component (fake override builder) that
generates fake overrides, computes their signatures and adds them to
the symbol table. Unlike the IR serializer/deserializer, it does not
know anything about indices in the IR file.

^KT-68992
^KT-72296