[Gradle] Don't add buildTypeAttribute for kotlin-android projects
After flipping flag `keepAndroidBuildTypeAttribute` to true by default
it broke kotlin-android publications that weren't configured on AGP side

In the previous commit we start reporting a warning for such projects
and here we make sure that BuildTypeAttribute is not set for kotlin
created components for single android projects.

This is a hack and we should stop creating/exposing
these kotlin components for kotlin-android projects.

^KT-72395
^KT-70380 Verification Pending