Fix incorrectly aggressive Unit coercion at elvis LHS
Initially expected type for LHS is brought at
8f2ad57f7a5d2a35a839b638575ec260900de439
to fix kt6694ExactAnnotationForElvis.kt.

At this point, the test works even without that expected type
 propagation.

But what's worse, this propagation in the case of Unit works incorrectly
because it forces LHS unit-coercion.

So, the solution is just resolving LHS in dependent mode as a regular
call argument.

Also, just to make RHS work consistently, we also update the mode
for it, which leads to quite a reasonable change
at elvisOnJavaList.fir.kt

Though the latter shouldn't change the result drastically because
elvis has @Exact on its return type, thus it would still ensure
the same constraint on RHS:
 Kv := ExpectedType
 RHSType <: Kv => RHSType <: ExpectedType

^KT-71751 Fixed