[Gradle] Add AndroidReleasePublicationIsNotConsumable diagnostic
After flipping default kotlin.android.buildTypeAttribute.keep to true
Users may get into troubles when their consumers depend on
android-release specific coordinates. Because of the presence of
BuildTypeAttribute.

The reason is because debug and release variants has to be present
in the root KMP publication. This is the key KMP feature that allows
users to declare dependencies via single GAV coordinates.

However presence of two different build type variants requires
KGP and AGP to specify BuildTypeAttribute explicitly.
So consumption of the Library is correct. However it has downsides that
BuildType attribute is also present in release publication and if some
AGP consumer tries to depend on it, AGP will fail with disambiguation
failure.

For example, imagine, user publishes library in the following layout:
* group:lib:1.0
* group:lib-jvm:1.0
* group:lib-android:1.0
* group:lib-android-debug:1.0

And there is android project that wants to consume the published library:
So it can depend on
`group:lib:1.0` -> release or debug will be selected depending on the
build type of their application.
`group:lib-jvm:1.0` -- JVM only implementation will be consumed
 it is not desirable, but possible.
 `group:lib-android-debug:1.0` -- debug artifact will be always selected
 `group:lib-android:1.0` -- unfortunately would fail

 Related gradle issue: https://github.com/gradle/gradle/issues/30642
 Related android issue: https://issuetracker.google.com/issues/374550337

 ^KT-72403 Verification Pending